pyproject.toml
[project]
name = "duck-contest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.35",
]
